Update exisitng CapiTemplates testdata yamls to GitOpsTemplate
