+++ test/utils/data/templates/cluster/aws/cluster-template-eks.yaml
@@ -1,8 +1,10 @@
-apiVersion: capi.weave.works/v1alpha1
-kind: CAPITemplate
+apiVersion: templates.weave.works/v1alpha1
+kind: GitOpsTemplate
 metadata:
   name: capa-cluster-template-eks-{{.Count}}
   namespace: default
+  labels:
+    weave.works/template-type: cluster
   annotations:    
     templates.weave.works/cost-estimation-enabled: "true"
     templates.weave.works/profiles-enabled: "false"
