+++ test/utils/data/templates/cluster/azure/cluster-template-e2e.yaml
@@ -1,9 +1,12 @@
-apiVersion: capi.weave.works/v1alpha1
-kind: CAPITemplate
+apiVersion: templates.weave.works/v1alpha1
+kind: GitOpsTemplate
 metadata:
   name: capz-cluster-template-{{.Count}}
   namespace: default
+  labels:
+    weave.works/template-type: cluster
   annotations:
+    templates.weave.works/inject-prune-annotation: "true"
     templates.weave.works/credentials-enabled: "true"
 spec:
   description: This is Azure capi quick start template-{{.Count}}
