+++ test/utils/data/templates/miscellaneous/invalid-cluster-template.yaml
@@ -1,8 +1,10 @@
-apiVersion: capi.weave.works/v1alpha1
-kind: CAPITemplate
+apiVersion: templates.weave.works/v1alpha1
+kind: GitOpsTemplate
 metadata:
   name: invalid-cluster-template-{{.Count}}
   namespace: default
+  labels:
+    weave.works/template-type: cluster
 spec:
   resourcetemplates:
     - apiVersion: cluster.x-k8s.io/v1alpha3
