+++ test/utils/data/tenancy/multiple-tenant.yaml
@@ -58,9 +58,6 @@ tenants:
         - apiGroups: ["cluster.x-k8s.io"]
           resources: ["clusters"]
           verbs: ["get", "list", "watch"]
-        - apiGroups: ["capi.weave.works"]
-          resources: ["capitemplates"]
-          verbs: ["get", "watch", "list"]
         - apiGroups: ["templates.weave.works"]
           resources: ["gitopstemplates"]
           verbs: ["get", "watch", "list"]
